Fixed PvP 99 gas caps. Fixed(?) RetreatSimple only checking safety 8px ahead. Made DTs more focused and cautious. PVP Citadels are now built sneakily with upgrade cancel. Zealot/Worker 99 micro should be better.

+++ src/Information/Battles/Types/JudgmentModifiers.scala
@@ -4,9 +4,9 @@ import Debugging.Visualizations.Colors
 import Lifecycle.With
 import Mathematics.Maff
 import Micro.Actions.Basic.Gather
-import Utilities.UnitFilters.{IsTank, IsWorker}
 import ProxyBwapi.Races.Terran
 import Utilities.Time.Minutes
+import Utilities.UnitFilters.{IsTank, IsWorker}
 import bwapi.Color
 
 import scala.collection.mutable.ArrayBuffer
@@ -39,7 +39,7 @@ object JudgmentModifiers {
     val keyBases        = With.geography.ourBasesAndSettlements.filter(b => b.isOurMain || b.isOurNatural)
     val distanceMax     = With.mapPixelWidth
     val distanceHome    = (if (keyBases.isEmpty) Seq(With.geography.home) else keyBases.map(_.heart.walkableTile)).map(centroid.groundPixels).min
-    val distanceRatio   = Maff.clamp(distanceHome.toDouble / distanceMax, 0, 1)
+    val distanceRatio   = Maff.clamp(distanceHome / distanceMax, 0, 1)
     val targetDeltaRaw  = (distanceRatio * 2 - 1) * 0.3
     val targetDelta     = Maff.clamp(targetDeltaRaw, deltaMin, deltaMax)
     Some(JudgmentModifier(targetDelta = targetDelta))
